test-skeleton: don't force -d on noMMU
There's a few tests whose do_test() is supposed to time out or terminate
with a signal. Tests communicate it to the test skeleton by defining
EXPECTED_SIGNAL or EXPECTED_STATUS and the skeleton completes with
success code when that happens. Running test suite in direct mode by
default changes that, as a result test running script reports failures
for the tests that actually pass or hangs.

Instead of switching to direct test running when building for noMMU use
vfork() and exec() the test in the child process in direct mode.
This makes the default test process behavior the same on MMU and noMMU
targets.

Signed-off-by: Max Filippov <jcmvbkbc@gmail.com>